+++ src/D20Tek.BlazorComponents.ResultValidator/ResultViewState.cs
@@ -1,0 +1,9 @@
+namespace D20Tek.BlazorComponents;
+
+public enum ResultViewState
+{
+    Loading,
+    Empty,
+    Success,
+    Failure
+}